# static/openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "license": {
      "name": "ISC"
    },
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).",
    "title": "",
    "version": ""
  },
  "servers": [
    {
      "url": "http://xolana.xen.network:5555"
    }
  ],
  "security": [
  ],
  "paths": {
...
    "/fetch_user/{pubkey}/{period}": {
      "get": {
        "summary": "/fetch_user/{pubkey}/{period}",
        "parameters": [
          {
            "name": "pubkey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        },
        "tags": []
      }
    },
...
    "/periods/{period}": {
      "get": {
        "summary": "/periods/{period}",
        "parameters": [
          {
            "name": "period",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "period": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "startBlockId": {
                            "type": "number"
                          },
                          "endBlockId": {
                            "type": "number"
                          },
                          "allocated": {
                            "type": "number"
                          },
                          "budget": {
                            "type": "number"
                          }
                        }
                      },
                      "example": [
                        {
                          "startBlockId": 29831401,
                          "endBlockId": 29917701,
                          "allocated": 1000.0274282579626,
                          "budget": 1000
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "tags": []
      }
    },
...
    "/voter/{pubkey}": {
      "get": {
        "summary": "/voter/{pubkey}",
        "parameters": [
          {
            "name": "pubkey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "periodNumber": {
                        "type": "number"
                      },
                      "reward": {
                        "type": "number"
                      },
                      "distributed": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "periodNumber",
                      "reward",
                      "distributed"
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": []
      }
    },
    "/voter/{pubkey}/balance": {
      "get": {
        "summary": "/voter/{pubkey}/balance",
        "parameters": [
          {
            "name": "pubkey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "voter": {
                      "type": "string",
                      "example": "25xW33GLtn1ukPg7g4d3H6vegKnsedvfjP3Fg6tMWHQA"
                    },
                    "accruedRewards": {
                      "type": "number",
                      "example": 25.721329381810826
                    },
                    "paidRewards": {
                      "type": "number",
                      "example": 0
                    },
                    "lastPeriod": {
                      "type": "number",
                      "example": 5
                    }
                  }
                }
              }
            }
          }
        },
        "tags": []
      }
    },
    "/voter/{pubkey}/payouts": {
      "get": {
        "summary": "/voter/{pubkey}/payouts",
        "parameters": [
          {
            "name": "pubkey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "example": 0
                }
              }
            }
          }
        },
        "tags": []
      }
    }
  },
...
}